NclexProject/requirements.txt
asgiref==3.9.1
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-extensions==4.1
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
tzdata==2025.2
